requirements.txt
numpy
sounddevice
matplotlib
pygame
pywebview